a2a_samples/pyproject.toml
[project]
name = "a2a-samples"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "a2a-sdk>=0.2.16",
    "asyncclick>=8.1.8",
    "httpx>=0.28.1",
    "uvicorn>=0.35.0",
]
